Make fallback to random(4) optional where getrandom(2) is available
The fallback to a character file is now gated on the `file-fallback`
feature (enabled by default); affects Linux/Android, macOS, and Solaris/Illumos.

Disabling this feature when targetting environments where getrandom(2) is
guaranteed to be present can be advantageous for reduced code size and ease of
audit/reasoning.

(Linux's random(4)/urandom(4) has silly blocking behaviour which makes it
difficult to implement sensible behaviour, and removing that code can be easier
than auditing it.)