Unconditionally use libc::getrandom on Illumos and Solaris
Also removes the use of `GRND_RANDOM`, which appears to be based on
outdated staements about the RNGs. For Solaris, see
[this blog post](https://blogs.oracle.com/solaris/post/solaris-new-system-calls-getentropy2-and-getrandom2).
For Illumos, the algorithms are less clear, but I don't see a clear
reason to continue using `GRND_RANDOM`.

I updated the documentation in `getrandom.rs` to full document this
decision and to have a common place listing when `getrandom(2)` became
avalible on each platform. I also updated the main lib.rs docs to point
to the correct man pages.

Note that Solaris 11.3 has a maximum buffer length of 1024 bytes,
while Illumos doesn't have this sort of issue.

Signed-off-by: Joe Richey <joerichey@google.com>