Rename `Weak` to `LazyPtr` and move it to `lazy.rs` (#427)
Currently the `Weak` type is used only by the NetBSD backend, but in
future it may be used by other backends. To simplify code a bit and
prepare for potential use on Windows, `Weak` now accepts a "pointer
initialization function" instead of a function name, i.e. it now works
similarly to `LazyUsize` and `LazyBool`.