Remove use of locale-specific `strerror_r`.
strerror_r doesn't necessarily return UTF-8 but we assume it does.

strerror_r is locale-sensitive but we are better off not being
locale-sensitive.

strerror_r requires us to use libc but increasingly we want to
avoid libc, e.g. to support x86_64-unknown-linux-none.

Just remove the use of the function.