Enforce OS errors are in the allowed range.
Avoid the `From<NonZeroU32>` implementation in favor of a constructor
that centralizes all the range checking in one place. Consistently use
 `Self::UNEXPECTED` for out-of-rnage values.

Besides being more consistent in the range checking, this also reduces
the boilerplate in callers, which makes it easier to maintain the ports
to less-common operating systems.

This is a step towards deprecating the `impl From<NonZeroU32> for Error`
implementation.