netbsd: Simplify weak lookup.
Remove LazyPtr. Avoid constructing an invalid pointer as a sentinel to
indicate that the pointer is uninitialized. Now, a null pointer means
it is uninitialized, and a non-null pointer means it is initialized.
This is less questionable from a safety perspective, and should also be
more efficient.

Reduce duplication between the "getrandom is available" and the
fallback case.