generators: Expose specific generators
Currently users of getrandom cannot choose which specific generator they
want to use if multiple are available on their system. As the preferred
generator may vary on a case-by-case basis, it is worth exposing an
interface that enables doing so.

Without affecting the preexisting functionality, this commit adds a new
"generators" module with a macro that can be used to generate function
wrappers with the same signature as getrandom() for each specific generator.

The generated functions can be used analogously to getrandom(), as follows:

getrandom(&mut buf).unwrap(); // <-- unchanged, default getrandom usage
generators::use_file(&mut buf).unwrap();
generators::linux_android_with_fallback(&mut buf).unwrap();

Due to the way conditional compilation of modules is currently handled
in the crate, this change required a rework of lib.rs in order to split
up selecting which modules to compile based on the target and selecting
which module to use as the preferred or default one.

In order to achieve that, a new module cfg_module.rs was added to
contain a macro cfg_if_module which takes a module name and a block of
code, then resolves to a cfg_if! containing the given block conditional
on the targets that the module can be compiled for. This way the targets
for each module can be configured once within the macro and then reused
across the crate as shown in the snippet below. This approach avoids
having to keep several sets of long target config attributes in sync in
different parts of the crate and makes the code more maintainable.

cfg_if_module!(linux_android_with_fallback, {
    mod linux_android_with_fallback;
});

This change does not intend to modify the default backend selection
logic. This split also makes it possible to avoid having to repeat the
same module declarations for modules that are dependencies of more than
one backend (especially mod util_libc).

As an example, an aarch64 Linux platform with the RNDR register available
could use the use_file, linux_android, linux_android_with_fallback, rndr
or rndr_with_fallback backends. All of them will be compiled and
accessible through "mod generators", but only one of them will be
selected and compiled as "mod imp" to be the crate default.