rndr: Add support for aarch64 RNDR register backend
AArch64 platforms from version Armv8.4 onwards may implement FEAT_RNG.
FEAT_RNG introduces the RNDR (and RNDRRS) register, reading from which
returns a random number.

Add support for using the RNDR register as a backend for getrandom.
The implementation is hidden behind a new "rndr" crate feature.
For the backend to work, users must ensure that the target platform
supports FEAT_RNG because it is not possible to reliably detect the
feature at runtime across platforms.