netbsd: fix fallback and test it using configuration flag (#555)
Fixes bug in the NetBSD fallback code based on `KERN_ARND`. `sysctl`
returns either 0 or -1, while the code assumed that on success it
returns number of bytes filled.

Adds the `getrandom_test_netbsd_fallback` configuration flag, which
forces the NetBSD backend to use the `KERN_ARND` fallback, and tweaks CI
to use the flag.