Web optimizations (#559)
As discussed in #541. This PR adds the following improvements:
- Caching of the global `Crypto` object.
- Detecting if our Wasm memory is based on a `SharedArrayBuffer`. If
not, we can copy bytes directly into our memory instead of having to go
through JS. This saves allocating the buffer in JS and copying the bytes
into Wasm memory. This is also the most common path. `SharedArrayBuffer`
requires `target_feature = "atomics"`, which is unstable and requires
Rust nightly. See
https://github.com/rust-random/getrandom/pull/559#issuecomment-2525342760
for full context.
- The atomic path only creates a sub-array when necessary, potentially
saving another FFI call.
- The atomic path will now allocate an `Uint8Array` with the minimum
amount of bytes necessary instead of a fixed size.
- The maximum chunk size for the non-atomic path and the maximum
`Uint8Array` size for the atomic paths have been increased to 65536
bytes: the maximum allowed buffer size for `Crypto.getRandomValues()`.

All in all this should give a performance improvement of ~5% to ~500%
depending on the amount of requested bytes and which path is taken. See
https://github.com/rust-random/getrandom/pull/559#discussion_r1874542668
for some benchmark results.

This spawned a bunch of improvements and fixes in `wasm-bindgen` that
are being used here:
- https://github.com/rustwasm/wasm-bindgen/pull/4315
- https://github.com/rustwasm/wasm-bindgen/pull/4316
- https://github.com/rustwasm/wasm-bindgen/pull/4318
- https://github.com/rustwasm/wasm-bindgen/pull/4319
- https://github.com/rustwasm/wasm-bindgen/pull/4340