Detect use of MemorySanitizer without using Nightly-only features (#571)
This allows msan detection to "just-work" whenever someone passes
`-Zsanitizer=memory`. Users no longer need to do any
`getrandom`-specific configuration.

This will also continue working once
https://github.com/rust-lang/rust/issues/123615 is merged, which
stabilizes some sanitizers (but not MemorySanitizer).

This is the approch taken by other low-level crates:
-
[`parking_lot_core`](https://github.com/Amanieu/parking_lot/blob/ca920b31312839013b4455aba1d53a4aede21b2f/core/build.rs)
-
[`crossbeam-utils`](https://github.com/crossbeam-rs/crossbeam/blob/00283fb1818174c25b02d7f1c883c5e19f8506a4/crossbeam-utils/build.rs#L42)

The only downside is that this adds a build-script, but it's as small as
possible, doesn't seem to impact build times, and is only a temporary
workaround.

---------

Signed-off-by: Joe Richey <joerichey@google.com>