windows: always use `RtlGenRandom` on pre-1.78 Rust versions (#610)
This PR adds detection of Rust compiler version in the build script
while compiling the crate for Windows targets.

If the version is smaller than 1.78, then we switch to the `windows7`
backend for all Windows targets. This is necessary because on those
older versions we can not disambiguate between win7 and win10 based
on the target triplet.