Add `linux_raw` opt-in backend (#572)
This PR replaces the `linux_rustix` opt-in backend introduced in #520
with `linux_raw` opt-in backend based raw syscalls implemented using
`asm!`. It noticeably reduces dependency footprint of the crate (see
concerns raised in #433). Unfortunately, it means that we have to
implement the raw syscalls ourselves.

This PR implements `linux_raw` support only for target arches with
stable `asm!`. Nightly-only support for other targets can be added in
later PRs.