Add `efi_rng` opt-in backend (#570)
The implementation is based on the [`std` code][0].
I am not sure why we need the try loop over protocol handles instead of
just getting the first handle, but I decided to just follow `std` here.

[0]: https://github.com/rust-lang/rust/blob/master/library/std/src/sys/random/uefi.rs

Because the backend requires access to `BootServices` it has to rely on
the unstable `uefi_std` feature. We could alternatively use the `uefi`
crate, but it would tie the backend to the `uefi` crate environment,
which is probably not desirable.

`TEMP_EFI_ERROR` is used as a temporary solution and will be fixed
in a separate PR.