Tweak `Error` representation (#614)
On UEFI targets `Error` is now represented as `NonZeroUsize` and on
other targets as `NonZeroI32`. In the former case OS errors have the
highest bit set to 1, while in the latter case OS errors are represented
as negative integers.

Additionally, this PR removes `Error::INTERNAL_START` and
`Error::CUSTOM_START` associated constants since the constants are no
longer relevant and the inner representation is an internal detail. This
is technically a breaking change, but I highly doubt that downstream
users rely on them since they can not be used for anything useful, so we
can consider it as a "fix" of unintentionally exported constants.