Implement Memory Sanitizer unpoisoning more precisely. (#678)
Avoid masking bugs in backends that return `Ok` without writing the
entire output buffer. For built-in backends, this makes MSAN useful for
validating the correctness of the control flow of each backend. For
custom backends in particular, this makes MSAN useful for validating
that a custom backend actually filled the entire output buffer when it
returned `Ok`.

When MSAN is enabled, this is a breaking change for a custom backend
that is implemented in a way that avoids built-in unpoisoning provided
by Rust/libc/MSAN.. Since MSAN support is unstable in Rust anyway, I
think this is acceptable breakage.

As a side effect, this minimizes the number of configurations that
reference the `__msan_unpoison` symbol unnecessarily.